Спринт 15. Реорганизация структуры проекта + добавление бронирований в dto предметов владельца
