--- src/main/java/ru/practicum/shareit/booking/BookingState.java
@@ -1,10 +1,0 @@
-package ru.practicum.shareit.booking;
-
-public enum BookingState {
-    ALL,
-    PAST,
-    CURRENT,
-    FUTURE,
-    WAITING,
-    REJECTED
-}
