--- src/main/java/ru/practicum/shareit/booking/BookingStatus.java
@@ -1,8 +1,0 @@
-package ru.practicum.shareit.booking;
-
-public enum BookingStatus {
-    WAITING,
-    APPROVED,
-    REJECTED,
-    CANCELLED
-}
