+++ src/main/java/ru/practicum/shareit/core/booking/BookingState.java
@@ -1,0 +1,10 @@
+package ru.practicum.shareit.core.booking;
+
+public enum BookingState {
+    ALL,
+    PAST,
+    CURRENT,
+    FUTURE,
+    WAITING,
+    REJECTED
+}
