+++ src/main/java/ru/practicum/shareit/core/booking/BookingStatus.java
@@ -1,0 +1,8 @@
+package ru.practicum.shareit.core.booking;
+
+public enum BookingStatus {
+    WAITING,
+    APPROVED,
+    REJECTED,
+    CANCELLED
+}
