+++ src/main/java/ru/practicum/shareit/core/item/persistance/entity/dto/ItemDto.java
@@ -1,0 +1,18 @@
+package ru.practicum.shareit.core.item.persistance.entity.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class ItemDto {
+    private Long id;
+    @NotBlank
+    private String name;
+    @NotBlank
+    private String description;
+    @NotNull
+    private Boolean available;
+}
