+++ src/main/java/ru/practicum/shareit/core/item/persistance/entity/dto/ItemDtoMapper.java
@@ -1,0 +1,61 @@
+package ru.practicum.shareit.core.item.persistance.entity.dto;
+
+import ru.practicum.shareit.core.booking.persistance.entity.model.Booking;
+import ru.practicum.shareit.core.booking.persistance.entity.dto.BookingDtoMapper;
+import ru.practicum.shareit.core.item.persistance.entity.model.Item;
+import ru.practicum.shareit.core.user.persistance.entity.model.User;
+
+import java.time.LocalDateTime;
+import java.util.Comparator;
+import java.util.List;
+import java.util.Optional;
+
+public class ItemDtoMapper {
+    private ItemDtoMapper() {
+
+    }
+
+    public static ItemDto toItemDto(Item item) {
+        return ItemDto.builder()
+                .id(item.getId())
+                .name(item.getName())
+                .description(item.getDescription())
+                .available(item.getAvailable())
+                .build();
+    }
+
+    public static Item toItem(ItemDto itemDto, User owner) {
+        return Item.builder()
+                .id(itemDto.getId())
+                .name(itemDto.getName())
+                .description(itemDto.getDescription())
+                .available(itemDto.getAvailable())
+                .owner(owner)
+                .build();
+    }
+
+    public static ItemOwnerDto toItemOwnerDto(Item item, List<Booking> bookings) {
+        ItemOwnerDto itemOwnerDto = ItemOwnerDto.builder()
+                .id(item.getId())
+                .available(item.getAvailable())
+                .name(item.getName())
+                .description(item.getDescription())
+                .build();
+
+        Optional<Booking> currentBooking = bookings.stream()
+                .filter(booking -> booking.getItem().getId().equals(item.getId()))
+                .filter(booking -> booking.getStart().isBefore(LocalDateTime.now()) && booking.getEnd().isAfter(LocalDateTime.now()))
+                .findFirst();
+
+        currentBooking.ifPresent(booking -> itemOwnerDto.setLastBooking(BookingDtoMapper.toBookingDto(booking)));
+
+        Optional<Booking> nextBooking = bookings.stream()
+                .filter(booking -> booking.getItem().getId().equals(item.getId()))
+                .filter(booking -> booking.getStart().isAfter(LocalDateTime.now()))
+                .min(Comparator.comparing(Booking::getStart));
+
+        nextBooking.ifPresent(booking -> itemOwnerDto.setNextBooking(BookingDtoMapper.toBookingDto(booking)));
+
+        return itemOwnerDto;
+    }
+}
