+++ src/main/java/ru/practicum/shareit/core/item/persistance/entity/model/Comment.java
@@ -1,0 +1,4 @@
+package ru.practicum.shareit.core.item.persistance.entity.model;
+
+public class Comment {
+}
