+++ src/main/java/ru/practicum/shareit/core/item/persistance/entity/model/Item.java
@@ -1,0 +1,35 @@
+package ru.practicum.shareit.core.item.persistance.entity.model;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import ru.practicum.shareit.core.user.persistance.entity.model.User;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "items")
+public class Item {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "name", nullable = false)
+    private String name;
+
+    @Column(name = "description", nullable = false)
+    private String description;
+
+    @Column(name = "available", nullable = false)
+    private Boolean available;
+
+    @ManyToOne
+    private User owner;
+
+    @Column(name = "request_id", nullable = false)
+    private Long request;
+}
