+++ src/main/java/ru/practicum/shareit/core/request/ItemRequestController.java
@@ -1,0 +1,12 @@
+package ru.practicum.shareit.core.request;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+/**
+ * TODO Sprint add-item-requests.
+ */
+@RestController
+@RequestMapping(path = "/requests")
+public class ItemRequestController {
+}
