+++ src/main/java/ru/practicum/shareit/core/request/dto/ItemRequest.java
@@ -1,0 +1,14 @@
+package ru.practicum.shareit.core.request.dto;
+
+import lombok.Data;
+
+import java.time.LocalDateTime;
+
+@Data
+public class ItemRequest {
+    private Long id;
+    private String description;
+    private Long requestor;
+    private LocalDateTime created;
+
+}
