+++ src/main/java/ru/practicum/shareit/core/request/dto/ItemRequestDto.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.core.request.dto;
+
+/**
+ * TODO Sprint add-item-requests.
+ */
+public class ItemRequestDto {
+}
