+++ src/main/java/ru/practicum/shareit/core/user/UserController.java
@@ -1,0 +1,36 @@
+package ru.practicum.shareit.core.user;
+
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.*;
+import ru.practicum.shareit.core.user.persistance.entity.dto.UserDto;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping(path = "/users")
+public class UserController {
+    private final UserService userService;
+
+    @GetMapping("/{id}")
+    public UserDto findById(@PathVariable Long id) {
+        return userService.findById(id);
+    }
+
+    @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
+    public UserDto create(@RequestBody @Valid UserDto user) {
+        return userService.create(user);
+    }
+
+    @PatchMapping("/{id}")
+    public UserDto update(@PathVariable Long id,
+                          @RequestBody UserDto user) {
+        return userService.update(id, user);
+    }
+
+    @DeleteMapping("/{id}")
+    public void delete(@PathVariable long id) {
+        userService.delete(id);
+    }
+}
