+++ src/main/java/ru/practicum/shareit/core/user/persistance/entity/dto/UserDtoMapper.java
@@ -1,0 +1,26 @@
+package ru.practicum.shareit.core.user.persistance.entity.dto;
+
+import ru.practicum.shareit.core.user.persistance.entity.model.User;
+
+public class UserDtoMapper {
+    private UserDtoMapper() {
+
+    }
+
+    public static UserDto toUserDto(User user) {
+        return UserDto.builder()
+                .id(user.getId())
+                .name(user.getName())
+                .email(user.getEmail())
+                .build();
+    }
+
+    public static User toUser(UserDto userDto) {
+        return User.builder()
+                .id(userDto.getId())
+                .name(userDto.getName())
+                .email(userDto.getEmail())
+                .build();
+    }
+
+}
