+++ src/main/java/ru/practicum/shareit/core/user/persistance/entity/model/User.java
@@ -1,0 +1,29 @@
+package ru.practicum.shareit.core.user.persistance.entity.model;
+
+import jakarta.persistence.*;
+import jakarta.validation.constraints.Email;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import static jakarta.persistence.GenerationType.IDENTITY;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "users")
+public class User {
+    @Id
+    @GeneratedValue(strategy = IDENTITY)
+    private Long id;
+
+    @Column(name = "name", nullable = false)
+    private String name;
+
+    @Email(message = "Email имеет некорректный формат")
+    @Column(name = "email", nullable = false)
+    private String email;
+}
