+++ src/main/java/ru/practicum/shareit/core/user/persistance/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package ru.practicum.shareit.core.user.persistance.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import ru.practicum.shareit.core.user.persistance.entity.model.User;
+
+import java.util.List;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    List<User> findAllByEmailAndIdNot(String email, Long id);
+}
