--- src/main/java/ru/practicum/shareit/item/persistance/entity/Item.java
@@ -1,35 +1,0 @@
-package ru.practicum.shareit.item.persistance.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import ru.practicum.shareit.user.persistance.entity.User;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "items")
-public class Item {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "name", nullable = false)
-    private String name;
-
-    @Column(name = "description", nullable = false)
-    private String description;
-
-    @Column(name = "available", nullable = false)
-    private Boolean available;
-
-    @ManyToOne
-    private User owner;
-
-    @Column(name = "request_id", nullable = false)
-    private Long request;
-}
