--- src/main/java/ru/practicum/shareit/item/persistance/entity/ItemDto.java
@@ -1,18 +1,0 @@
-package ru.practicum.shareit.item.persistance.entity;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class ItemDto {
-    private Long id;
-    @NotBlank
-    private String name;
-    @NotBlank
-    private String description;
-    @NotNull
-    private Boolean available;
-}
