--- src/main/java/ru/practicum/shareit/item/persistance/entity/ItemDtoMapper.java
@@ -1,28 +1,0 @@
-package ru.practicum.shareit.item.persistance.entity;
-
-import ru.practicum.shareit.user.persistance.entity.User;
-
-public class ItemDtoMapper {
-    private ItemDtoMapper() {
-
-    }
-
-    public static ItemDto toItemDto(Item item) {
-        return ItemDto.builder()
-                .id(item.getId())
-                .name(item.getName())
-                .description(item.getDescription())
-                .available(item.getAvailable())
-                .build();
-    }
-
-    public static Item toItem(ItemDto itemDto, User owner) {
-        return Item.builder()
-                .id(itemDto.getId())
-                .name(itemDto.getName())
-                .description(itemDto.getDescription())
-                .available(itemDto.getAvailable())
-                .owner(owner)
-                .build();
-    }
-}
