--- src/main/java/ru/practicum/shareit/item/persistance/repository/ItemRepository.java
@@ -1,18 +1,0 @@
-package ru.practicum.shareit.item.persistance.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.stereotype.Repository;
-import ru.practicum.shareit.item.persistance.entity.Item;
-
-import java.util.List;
-
-@Repository
-public interface ItemRepository extends JpaRepository<Item, Long> {
-    List<Item> findAllByOwnerId(Long ownerId);
-
-    @Query("select item from Item item " +
-            "where item.available = TRUE and (upper(item.name) like upper(concat('%', ?1, '%')) " +
-            "or upper(item.description) like upper(concat('%', ?1, '%')))")
-    List<Item> findAllBySearch(String search);
-}
