--- src/main/java/ru/practicum/shareit/request/ItemRequestController.java
@@ -1,12 +1,0 @@
-package ru.practicum.shareit.request;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-/**
- * TODO Sprint add-item-requests.
- */
-@RestController
-@RequestMapping(path = "/requests")
-public class ItemRequestController {
-}
