--- src/main/java/ru/practicum/shareit/request/dto/ItemRequest.java
@@ -1,14 +1,0 @@
-package ru.practicum.shareit.request.dto;
-
-import lombok.Data;
-
-import java.time.LocalDateTime;
-
-@Data
-public class ItemRequest {
-    private Long id;
-    private String description;
-    private Long requestor;
-    private LocalDateTime created;
-
-}
