--- src/main/java/ru/practicum/shareit/user/persistance/entity/UserDto.java
@@ -1,16 +1,0 @@
-package ru.practicum.shareit.user.persistance.entity;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class UserDto {
-    private Long id;
-    private String name;
-    @NotBlank(message = "Email не должен быть пустым")
-    @Email(message = "Email имеет некорректный формат")
-    private String email;
-}
