--- src/main/java/ru/practicum/shareit/user/persistance/entity/UserDtoMapper.java
@@ -1,24 +1,0 @@
-package ru.practicum.shareit.user.persistance.entity;
-
-public class UserDtoMapper {
-    private UserDtoMapper() {
-
-    }
-
-    public static UserDto toUserDto(User user) {
-        return UserDto.builder()
-                .id(user.getId())
-                .name(user.getName())
-                .email(user.getEmail())
-                .build();
-    }
-
-    public static User toUser(UserDto userDto) {
-        return User.builder()
-                .id(userDto.getId())
-                .name(userDto.getName())
-                .email(userDto.getEmail())
-                .build();
-    }
-
-}
