--- src/main/java/ru/practicum/shareit/user/persistance/repository/UserRepository.java
@@ -1,12 +1,0 @@
-package ru.practicum.shareit.user.persistance.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import ru.practicum.shareit.user.persistance.entity.User;
-
-import java.util.List;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    List<User> findAllByEmailAndIdNot(String email, Long id);
-}
