+++ src/test/java/ru/practicum/shareit/item/ItemControllerTest.java
@@ -5,11 +5,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
 import ru.practicum.shareit.ShareItApp;
+import ru.practicum.shareit.core.item.ItemController;
 import ru.practicum.shareit.exception.ConditionsNotMetException;
 import ru.practicum.shareit.exception.NotFoundException;
-import ru.practicum.shareit.item.persistance.entity.ItemDto;
-import ru.practicum.shareit.user.UserController;
-import ru.practicum.shareit.user.persistance.entity.UserDto;
+import ru.practicum.shareit.core.item.persistance.entity.dto.ItemDto;
+import ru.practicum.shareit.core.user.UserController;
+import ru.practicum.shareit.core.user.persistance.entity.dto.UserDto;
 
 import static org.junit.jupiter.api.Assertions.*;
